Introduced a more basic version of AbstractRenderLayer and changed the previous AbstractRenderLayer to AbstractOrderableRenderLayer.
