+++ Artemis.Engine/Artemis.Engine.csproj
@@ -54,6 +54,7 @@
     <Compile Include="Form.cs" />
     <Compile Include="GameConstants.cs" />
     <Compile Include="GameSetupException.cs" />
+    <Compile Include="Graphics\AbstractOrderableRenderLayer.cs" />
     <Compile Include="Graphics\AbstractRenderLayer.cs" />
     <Compile Include="Graphics\Animation\AAMLException.cs" />
     <Compile Include="Graphics\Animation\AAMLFileReader.cs" />
